% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod_growth.R
\name{cod_growth}
\alias{cod_growth}
\title{cod_growth}
\usage{
cod_growth(a, b, c, d)
}
\arguments{
\item{a}{first variable in best fit third order polynomial of cod fish growth (for more details see references)}

\item{b}{second variable in best fit third order polynomial of cod fish growth (for more details see references)}

\item{c}{third variable in best fit third order polynomial of cod fish growth (for more details see references)}

\item{d}{first variable in best fit third order polynomial of cod fish growth (for more details see references)}
}
\value{
Estimated growth rate for cod from the temperature range 0-20 degrees C
}
\description{
Function to estimate cod growth at a given temperature (C)
}
\references{
[doi:10.1016/j.aquaculture.2007.06.026]
}
\author{
Nicol Parker
}
