% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_catch.R
\name{fish_catch}
\alias{fish_catch}
\title{fish_catch}
\usage{
fish_catch(P, C, G)
}
\arguments{
\item{P}{prices of fish species of interest}

\item{C}{number of catch of each fish species at a given location}
}
\description{
Function to evaluate wild caught fish species, price, and revenue in Lake Michigan and Lake Superior
}
\author{
Nicol Parker
}
